Add gate function checks to controller methods

+++ app/Http/Controllers/Admin/OrganizationController.php
@@ -10,6 +10,7 @@ use Inertia\Inertia;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Http\RedirectResponse;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 
 class OrganizationController extends Controller
 {
@@ -18,11 +19,14 @@ class OrganizationController extends Controller
      */
     public function index()
     {
-        $organizations = Organization::all();
+        if (Gate::allows('view_organization')) {
+            $organizations = Organization::all();
 
-        return Inertia::render('Admin/Organizations/Index', [
-            'organizations' => $organizations
-        ]);
+            return Inertia::render('Admin/Organizations/Index', [
+                'organizations' => $organizations
+            ]);
+        }
+        return redirect('/');
     }
 
     /**
@@ -30,11 +34,14 @@ class OrganizationController extends Controller
      */
     public function create()
     {
-        $users = User::all();
+        if (Gate::allows('create_organization')) {
+            $users = User::all();
 
-        return Inertia::render('Admin/Organizations/Create', [
-            'users' => $users
-        ]);
+            return Inertia::render('Admin/Organizations/Create', [
+                'users' => $users
+            ]);
+        }
+        return redirect('/');
     }
 
     /**
@@ -42,54 +49,57 @@ class OrganizationController extends Controller
      */
     public function store(Request $request)
     {
-        $validated = $request->validate([
-            'slug' => 'required|max:255|unique:organizations,slug', // 'slug' is a unique field in the 'organizations' table
-            'owner' => 'required|exists:users,id',
-            'name' => 'required|max:255',
-            'logo' => 'nullable|image',
-            'cover' => 'nullable|image',
-            'description' => 'required|max:1024',
-            'address_line_1' => 'required|max:255',
-            'address_line_2' => 'nullable|max:255',
-            'district' => 'required|max:255',
-            'country' => 'required|max:255',
-            'postal_code' => 'required|max:255',
-            'email' => 'required|email|unique:organizations,email',
-            'number' => 'nullable|max:255',
-            'whatsapp' => 'nullable|max:255',
-            'whatsapp_group' => 'nullable|max:255',
-            'facebook' => 'nullable|max:255',
-            'instagram' => 'nullable|max:255',
-            'twitter' => 'nullable|max:255',
-            'website' => 'nullable|max:255',
-            'youtube' => 'nullable|max:255',
-            'linkedin' => 'nullable|max:255',
-            'is_active' => 'required|boolean',
-            'is_verified' => 'required|boolean',
-            'created_by' => 'required|exists:users,id',
-        ]);
+        if (Gate::allows('create_organization')) {
+            $validated = $request->validate([
+                'slug' => 'required|max:255|unique:organizations,slug', // 'slug' is a unique field in the 'organizations' table
+                'owner' => 'required|exists:users,id',
+                'name' => 'required|max:255',
+                'logo' => 'nullable|image',
+                'cover' => 'nullable|image',
+                'description' => 'required|max:1024',
+                'address_line_1' => 'required|max:255',
+                'address_line_2' => 'nullable|max:255',
+                'district' => 'required|max:255',
+                'country' => 'required|max:255',
+                'postal_code' => 'required|max:255',
+                'email' => 'required|email|unique:organizations,email',
+                'number' => 'nullable|max:255',
+                'whatsapp' => 'nullable|max:255',
+                'whatsapp_group' => 'nullable|max:255',
+                'facebook' => 'nullable|max:255',
+                'instagram' => 'nullable|max:255',
+                'twitter' => 'nullable|max:255',
+                'website' => 'nullable|max:255',
+                'youtube' => 'nullable|max:255',
+                'linkedin' => 'nullable|max:255',
+                'is_active' => 'required|boolean',
+                'is_verified' => 'required|boolean',
+                'created_by' => 'required|exists:users,id',
+            ]);
 
-        $organization = new Organization();
-        $organization->fill($validated);
+            $organization = new Organization();
+            $organization->fill($validated);
 
-        if ($request->hasFile('logo')) {
-            $logo = $request->file('logo');
-            $filename = $request->name . '-logo-' . time() . '.' . $logo->getClientOriginalExtension();
-            $logo->storeAs('public/images/org/logo/', $filename);
-            $organization->logo = $filename;
+            if ($request->hasFile('logo')) {
+                $logo = $request->file('logo');
+                $filename = $request->name . '-logo-' . time() . '.' . $logo->getClientOriginalExtension();
+                $logo->storeAs('public/images/org/logo/', $filename);
+                $organization->logo = $filename;
+            }
+
+            if ($request->hasFile('cover')) {
+                $cover = $request->file('cover');
+                $coverName = $request->name . '-cover-' . time() . '.' . $cover->getClientOriginalExtension();
+                $cover->storeAs('public/images/org/cover/', $coverName);
+                $organization->cover = $coverName;
+            }
+
+            $organization->save();
+
+
+            return Redirect::route('organizations.index')->with('success', 'Organization has been created');
         }
-
-        if ($request->hasFile('cover')) {
-            $cover = $request->file('cover');
-            $coverName = $request->name . '-cover-' . time() . '.' . $cover->getClientOriginalExtension();
-            $cover->storeAs('public/images/org/cover/', $coverName);
-            $organization->cover = $coverName;
-        }
-
-        $organization->save();
-
-
-        return Redirect::route('organizations.index')->with('success', 'Organization has been created');
+        return redirect('/');
     }
 
     /**
@@ -97,11 +107,14 @@ class OrganizationController extends Controller
      */
     public function show(string $id)
     {
-        $organization = Organization::findOrFail($id);
+        if (Gate::allows('view_organization')) {
+            $organization = Organization::findOrFail($id);
 
-        return Inertia::render('Admin/Organizations/Show', [
-            'organization' => $organization
-        ]);
+            return Inertia::render('Admin/Organizations/Show', [
+                'organization' => $organization
+            ]);
+        }
+        return redirect('/');
     }
 
     /**
@@ -109,13 +122,16 @@ class OrganizationController extends Controller
      */
     public function edit(string $id)
     {
-        $organization = Organization::findOrFail($id);
-        $users = User::all();
+        if (Gate::allows('update_organization')) {
+            $organization = Organization::findOrFail($id);
+            $users = User::all();
 
-        return Inertia::render('Admin/Organizations/Edit', [
-            'organization' => $organization,
-            'users' => $users
-        ]);
+            return Inertia::render('Admin/Organizations/Edit', [
+                'organization' => $organization,
+                'users' => $users
+            ]);
+        }
+        return redirect('/');
     }
 
     /**
@@ -123,40 +139,43 @@ class OrganizationController extends Controller
      */
     public function update(Request $request, string $id): RedirectResponse
     {
-        $validated = $request->validate([
-            'slug' => 'required|max:255|unique:organizations,slug,' . $id, // 'slug' is a unique field in the 'organizations' table
-            'owner' => 'required|exists:users,id',
-            'name' => 'required|max:255',
-            'description' => 'required|max:1024',
-            'address_line_1' => 'required|max:255',
-            'address_line_2' => 'nullable|max:255',
-            'district' => 'required|max:255',
-            'country' => 'required|max:255',
-            'postal_code' => 'required|max:255',
-            'number' => 'nullable|max:255',
-            'whatsapp' => 'nullable|max:255',
-            'whatsapp_group' => 'nullable|max:255',
-            'facebook' => 'nullable|max:255',
-            'instagram' => 'nullable|max:255',
-            'twitter' => 'nullable|max:255',
-            'website' => 'nullable|max:255',
-            'youtube' => 'nullable|max:255',
-            'linkedin' => 'nullable|max:255',
-            'is_active' => 'required|boolean',
-            'is_verified' => 'required|boolean',
-        ]);
+        if (Gate::allows('update_organization')) {
+            $validated = $request->validate([
+                'slug' => 'required|max:255|unique:organizations,slug,' . $id, // 'slug' is a unique field in the 'organizations' table
+                'owner' => 'required|exists:users,id',
+                'name' => 'required|max:255',
+                'description' => 'required|max:1024',
+                'address_line_1' => 'required|max:255',
+                'address_line_2' => 'nullable|max:255',
+                'district' => 'required|max:255',
+                'country' => 'required|max:255',
+                'postal_code' => 'required|max:255',
+                'number' => 'nullable|max:255',
+                'whatsapp' => 'nullable|max:255',
+                'whatsapp_group' => 'nullable|max:255',
+                'facebook' => 'nullable|max:255',
+                'instagram' => 'nullable|max:255',
+                'twitter' => 'nullable|max:255',
+                'website' => 'nullable|max:255',
+                'youtube' => 'nullable|max:255',
+                'linkedin' => 'nullable|max:255',
+                'is_active' => 'required|boolean',
+                'is_verified' => 'required|boolean',
+            ]);
 
-        if ($id) {
-            $organization = Organization::findOrFail($id);
-        } else {
-            $organization = new Organization();
+            if ($id) {
+                $organization = Organization::findOrFail($id);
+            } else {
+                $organization = new Organization();
+            }
+
+            $organization->fill($validated);
+
+            $organization->save();
+
+            return Redirect::route('organizations.index')->with('update', 'Organization has been updated');
         }
-
-        $organization->fill($validated);
-
-        $organization->save();
-
-        return Redirect::route('organizations.index')->with('update', 'Organization has been updated');
+        return redirect('/');
     }
 
     /**
@@ -164,20 +183,23 @@ class OrganizationController extends Controller
      */
     public function destroy(string $id)
     {
-        $organization = Organization::findOrFail($id);
+        if (Gate::allows('delete_organization')) {
+            $organization = Organization::findOrFail($id);
 
-        // Delete the logo file if it exists
-        if ($organization->logo) {
-            Storage::delete('public/images/org/logo/' . $organization->logo);
+            // Delete the logo file if it exists
+            if ($organization->logo) {
+                Storage::delete('public/images/org/logo/' . $organization->logo);
+            }
+
+            // Delete the cover file if it exists
+            if ($organization->cover) {
+                Storage::delete('public/images/org/cover/' . $organization->cover);
+            }
+
+            $organization->delete();
+
+            return Redirect::route('organizations.index')->with('delete', 'Organization has been deleted');
         }
-
-        // Delete the cover file if it exists
-        if ($organization->cover) {
-            Storage::delete('public/images/org/cover/' . $organization->cover);
-        }
-
-        $organization->delete();
-
-        return Redirect::route('organizations.index')->with('delete', 'Organization has been deleted');
+        return redirect('/');
     }
 }
